+++ example/djangoexample/views.py
@@ -27,10 +27,6 @@ def make_child2_queryset() -> DefaultQuerySet[Child2]:
     return Child2.objects.all()
 
 
-def make_child_typevar_queryset(child: type[T_Child]) -> DefaultQuerySet[T_Child]:
-    return child.objects.all()
-
-
 def ones(model: type[Concrete[Parent]]) -> list[str]:
     # Union[django.db.models.manager.Manager[djangoexample.exampleapp.models.Child1], djangoexample.exampleapp.models.ManagerFromChild2QuerySet[djangoexample.exampleapp.models.Child2], django.db.models.manager.Manager[djangoexample.exampleapp.models.Child3]]
     reveal_type(model.objects)
@@ -59,13 +55,5 @@ def index(request: HttpRequest) -> HttpResponseBase:
     reveal_type(Child2.objects)
     # djangoexample.exampleapp.models.Child2QuerySet[djangoexample.exampleapp.models.Child2]
     reveal_type(Child2.objects.all())
-
-    tvqs1 = make_child_typevar_queryset(Child1)
-    # django.db.models.query._QuerySet[djangoexample.exampleapp.models.Child1]
-    reveal_type(tvqs1)
-
-    tvqs2 = make_child_typevar_queryset(Child2)
-    # djangoexample.exampleapp.models.Child2QuerySet
-    reveal_type(tvqs2)
 
     return HttpResponse("Hello there")