+++ tests/test_works.py
@@ -38,10 +38,6 @@ def test_works(scenario: Scenario) -> None:
                 return child.objects.all()
 
 
-            def make_child_typevar_queryset(child: type[T_Child]) -> DefaultQuerySet[T_Child]:
-                return child.objects.all()
-
-
             def ones(model: type[Concrete[Parent]]) -> list[str]:
                 model.objects
                 # ^ REVEAL ^ Union[django.db.models.manager.Manager[myapp.models.Child1], myapp.models.ManagerFromChild2QuerySet[myapp.models.Child2], django.db.models.manager.Manager[myapp.models.Child3], django.db.models.manager.Manager[myapp2.models.ChildOther]]
@@ -68,12 +64,6 @@ def test_works(scenario: Scenario) -> None:
 
             Child2.objects.all()
             # ^ REVEAL ^ myapp.models.Child2QuerySet[myapp.models.Child2]
-
-            make_child_typevar_queryset(Child1)
-            # ^ REVEAL ^ django.db.models.query.QuerySet[myapp.models.Child1, myapp.models.Child1]
-
-            make_child_typevar_queryset(Child2)
-            # ^ REVEAL ^ myapp.models.Child2QuerySet
 
             make_multiple_queryset(Child1)
             # ^ REVEAL ^ Union[myapp.models.Child2QuerySet, django.db.models.query.QuerySet[myapp.models.Child1, myapp.models.Child1]]