+++ docs/api/changelog.rst
@@ -16,6 +16,9 @@ Changelog
       variable name as the result.
     * Removed Concrete.type_var. Can use
       ``TypeVar("T_Model", bound=Concrete[Model])`` instead
+    * Previous versions would create an implicit ``# type: ignore[return-value]`` on the
+      return statement of methods/functions that returned a concrete annotation of a
+      type var. This is no longer the case.
 
 .. _release-0.6.4: